fix: renamed StakeReview to StakeConfirmation
