+++ app/components/UI/Stake/Views/StakeConfirmationView/StakeConfirmationMockData.ts
