+++ app/components/UI/Stake/Views/StakeConfirmationView/StakeConfirmationView.styles.ts
