+++ app/components/UI/Stake/Views/StakeConfirmationView/StakeConfirmationView.types.ts
@@ -1,0 +1,10 @@
+import { RouteProp } from '@react-navigation/native';
+
+interface StakeConfirmationViewRouteParams {
+  wei: string;
+  fiat: string;
+}
+
+export interface StakeConfirmationViewProps {
+  route: RouteProp<{ params: StakeConfirmationViewRouteParams }, 'params'>;
+}
