--- app/components/UI/Stake/Views/StakeReviewView/StakeReviewView.types.ts
@@ -1,10 +1,0 @@
-import { RouteProp } from '@react-navigation/native';
-
-interface StakeReviewViewParams {
-  wei: string;
-  fiat: string;
-}
-
-export interface StakeReviewViewProps {
-  route: RouteProp<{ params: StakeReviewViewParams }, 'params'>;
-}
