+++ app/constants/navigation/Routes.ts
@@ -139,7 +139,7 @@ const Routes = {
   },
   STAKING: {
     STAKE: 'Stake',
-    STAKE_REVIEW: 'StakeReview',
+    STAKE_CONFIRMATION: 'StakeConfirmation',
     UNSTAKE: 'Unstake',
     CLAIM: 'Claim',
     MODALS: {
